+++ Shared/Libs/Interfaces/IDrawable.hpp
@@ -14,8 +14,8 @@ public:
 
 public:
     virtual void Draw(sf::RenderTexture &rect) = 0;
-    virtual vec2<int> GetRenderRect() = 0;
-    virtual vec2<int> GetPosition() = 0;
+    virtual vec2<float> GetRenderRect() = 0;
+    virtual vec2<float> GetPosition() = 0;
 };
 
 #endif //R_TYPE_IDRAWABLE_HPP
